BYIN-10916 Handle early crashes gracefully.
When early app crash happens at the moment of instrumentation
command starts Android instrumenation framework cannot
handle this and DDMS doesn't call "testStarted()" callback
and our 'failLastTest()' handler crashes with NPE.
To fix this we provide id of the first instremented test to
the 'failLastTest()' handler.
If we cannot do this we crash it with RuntimeException().